# [chore] Move to goccy-json