--- a/internal/splunk/common_test.go
+++ b/internal/splunk/common_test.go
@@ -4,10 +4,10 @@
 package splunk
 
 import (
-	"encoding/json"
 	"strings"
 	"testing"
 
+	"github.com/goccy/go-json"
 	"github.com/stretchr/testify/assert"
 )