… exporter/sapmexporter/go.sum
… exporter/signalfxexporter/go.sum
… internal/splunk/go.sum
… receiver/signalfxreceiver/go.sum