<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Moves the splunkhecreceiver to goccy-json from jsoniter.

This speeds up execution:


Before:
```
/Users/atoulme/sdk/go1.23.9/bin/go test -v ./... -bench . -run ^$ #gosetup
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver
cpu: Apple M4 Pro
BenchmarkHandleReq
BenchmarkHandleReq-12    	    5215	    413247 ns/op	  168600 B/op	    4403 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver	1.610s
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver/internal/metadata	0.210s

Process finished with the exit code 0
```

After:
```
/Users/atoulme/sdk/go1.23.9/bin/go test -v ./... -bench . -run ^$ #gosetup
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver
cpu: Apple M4 Pro
BenchmarkHandleReq
BenchmarkHandleReq-12    	    9963	    114607 ns/op	  169058 B/op	    2467 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver	2.360s
PASS
ok  	github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver/internal/metadata	0.190s

Process finished with the exit code 0

```

Benchstat:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver
cpu: Apple M4 Pro
             │ /Users/atoulme/Downloads/before.txt │ /Users/atoulme/Downloads/after.txt │
             │               sec/op                │     sec/op       vs base           │
HandleReq-12                          413.2µ ± ∞ ¹      114.6µ ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

             │ /Users/atoulme/Downloads/before.txt │ /Users/atoulme/Downloads/after.txt │
             │                B/op                 │      B/op        vs base           │
HandleReq-12                         164.6Ki ± ∞ ¹     165.1Ki ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

             │ /Users/atoulme/Downloads/before.txt │ /Users/atoulme/Downloads/after.txt │
             │              allocs/op              │    allocs/op     vs base           │
HandleReq-12                          4.403k ± ∞ ¹      2.467k ± ∞ ¹  ~ (p=1.000 n=1) ²
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

```